+++ src/main/java/team_questio/questio/portfolio/presentation/PortfolioController.java
@@ -1,4 +1,4 @@
-package team_questio.questio.portfolio.application;
+package team_questio.questio.portfolio.presentation;
 
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import team_questio.questio.portfolio.application.dto.PortfolioRequest;
-import team_questio.questio.portfolio.application.dto.PortfolioResponse;
-import team_questio.questio.portfolio.service.PortfolioFacadeService;
+import team_questio.questio.portfolio.presentation.dto.PortfolioRequest;
+import team_questio.questio.portfolio.presentation.dto.PortfolioResponse;
+import team_questio.questio.portfolio.application.PortfolioFacadeService;
 
 @Slf4j
 @RestController
