+++ src/main/java/team_questio/questio/portfolio/presentation/dto/PortfolioDetailResponse.java
@@ -1,6 +1,6 @@
-package team_questio.questio.portfolio.application.dto;
+package team_questio.questio.portfolio.presentation.dto;
 
-import team_questio.questio.portfolio.service.dto.PortfolioDetailInfo;
+import team_questio.questio.portfolio.application.dto.PortfolioDetailInfo;
 
 public record PortfolioDetailResponse(
         Long portfolioId,
