+++ src/main/java/team_questio/questio/portfolio/presentation/dto/PortfolioRequest.java
@@ -1,6 +1,6 @@
-package team_questio.questio.portfolio.application.dto;
+package team_questio.questio.portfolio.presentation.dto;
 
-import team_questio.questio.portfolio.service.dto.PortfolioParam;
+import team_questio.questio.portfolio.application.dto.PortfolioParam;
 
 public record PortfolioRequest(
         String content
